requirements.txt
asgiref==3.2.10
boto3==1.14.23
botocore==1.17.23
Django==3.0.8
docutils==0.15.2
jmespath==0.10.0
mysqlclient==2.0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
